adding more api types
